mise à jour de user et produit

+++ routes/produitRouter.js
@@ -24,7 +24,7 @@ produitRouter.route( '/' )
     } )
     .get( cors.corsWithOptions,  (req, res, next) => {
         Produit.find( {} )
-        .populate(['reseau', 'photos', 'ville', 'utilisateur'])
+        .populate(['user'])
         .then( ( produit ) =>{
             res.json( produit )
             }, ( err ) => { next( err ) 
@@ -33,13 +33,16 @@ produitRouter.route( '/' )
     })
     .post( cors.corsWithOptions, authenticate.verifyUser, upload.array('images', 12), ( req, res, next ) =>{
         req.body.images = [];
-        req.body.ville = mongoose.Types.ObjectId( req.body.ville.trim() );
-        req.body.user = mongoose.Types.ObjectId( req.body.user.trim() );
-        req.files.forEach( file => {
-            req.body.images.push(file.path.substr( 'public'.length ));
-        } );
-        
-        if( req.body.designation != undefined ) {
+        if( req.body.ville != undefined )
+            req.body.ville = mongoose.Types.ObjectId( req.body.ville.trim() );
+        if( req.body.user != undefined )
+            req.body.user = mongoose.Types.ObjectId( req.body.user.trim() );
+        if( req.files != undefined ) {
+            req.files.forEach( file => {
+                req.body.images.push(file.path.substr( 'public'.length ));
+            } );
+        }
+        if( req.body != undefined ) {
             Produit.create( req.body )
                 .then(( produit ) => {
                     console.log( 'produit crée ', produit );
@@ -74,7 +77,7 @@ produitRouter.route( '/:id' )
     } )
     .get( cors.cors, (req, res, next) => {
         Produit.findById( req.params.id )
-            .populate(['reseau', 'photos', 'ville', 'utilisateur'])
+            .populate([ 'user'])
             .then( ( produit ) => {
                 res.json( produit );
             }, (err) => next( err ) )
@@ -88,38 +91,65 @@ produitRouter.route( '/:id' )
         res.statusCode = 403;
         res.json({'message': 'La methode POST n\'est pas supporté.', 'status': res.statusCode } )
     } )
-    .put( cors.corsWithOptions, authenticate.verifyUser, ( req, res, next ) => {
-        //req.body.images = [];
-        if( req.body.images != null )
-            req.body.images = [];
-        if( req.body.ville !== null || req.body.ville == '' )
-            req.body.ville = mongoose.Types.ObjectId( req.body.ville.trim() );
-        if( Number(req.body.prix) > 0 )
-          
-        if( req.files.length > 0 || req.files != null ) {
-            req.files.forEach( file => {
-                req.body.images.push('/images/produits/' + file.originalname);
-            } );
-        }
+    .put( cors.corsWithOptions, authenticate.verifyUser, authenticate.productOwner, upload.array('images', 12), ( req, res, next ) => {
+        Produit.findById( req.params.id )
+            .then( ( produit ) => {
+            
+                if( req.body.user != undefined )
+                    req.body.user = mongoose.Types.ObjectId( req.body.user.trim() );
 
-        Produit.findByIdAndUpdate( req.params.id, {
-            $set: req.body
-        }, { new: true })
-        .then( ( ville ) => {
-            res.json( ville );
-        }, ( err ) => next( err ) )
-        .catch( ( err ) => next( err ) );
-    } )
-    .delete( cors.corsWithOptions, authenticate.verifyUser, ( req, res, next ) => {
-        Produit.findByIdAndRemove(req.params.id)
-            .then( ( resp ) => {
-                res.json(resp);
-            }, ( err ) => {
+                req.body.images = [];
+
+
+                if( req.files.length > 0 || req.files != null ) {
+                    req.files.forEach( file => {
+                        req.body.images.push('/images/produits/' + file.originalname);
+                    } );
+                }
+
+            if( produit.user.toString() === req.user._id.toString() ) {
+                    Produit.findOneAndUpdate( {_id: req.params.id} , {
+                        $set: req.body
+                    }, { new: true })
+                    .then( ( ville ) => {
+                        res.json( ville );
+                    }, ( err ) => next( err ) )
+                    .catch( ( err ) => next( err ) );
+            } else {
+                let err = new Error( 'Vous ne pouvez pas mettre à jour un ' + req.body.designation + ', ça ne vous appartient pas ?' );
+                err.status = 403;
+                res.statusCode = 403;
+                res.json( {statusCode: res.statusCode, statusText: err.message } );
+                }
+        }, (err) => next( err ) )
+        .catch( ( err ) => {
                 err = new Error('Produit ' + req.params.id + ' non trouvé'); next( err );
                 err.status = 404;
                 res.json({'message': err.message, 'status': err.status });
-            } )
-            .catch( ( err ) => next( err ) );
+        } );
+    } )
+    .delete( cors.corsWithOptions, authenticate.verifyUser, ( req, res, next ) => {
+        Produit.findById( req.params.id )
+        .then( ( produit ) => {
+            if( produit.user.toString() === req.user._id.toString() ) {
+                console.log( produit.user +'=='+ req.user._id)
+                Produit.findByIdAndRemove(req.params.id)
+                    .then( ( resp ) => {
+                        res.json(resp);
+                    }, ( err ) => {
+                        err = new Error('Produit ' + req.params.id + ' non trouvé'); next( err );
+                        err.status = 404;
+                        res.json({'message': err.message, 'status': err.status });
+                    } )
+                    .catch( ( err ) => next( err ) );
+            } else {
+                let err = new Error( 'Vous ne pouvez pas supprimé ' + produit.designation + ', ça ne vous appartient pas ?' );
+                err.status = 403;
+                res.statusCode = 403;
+                res.json( {statusCode: res.statusCode, statusText: err.message } );
+            } 
+        }, err => { next( err ) } );
+        
     } );
 
 module.exports = produitRouter;
